Remove hand smelling animation
